moved full_treatment_train model_components
