+++ watertap/examples/flowsheets/full_treatment_train/model_components/property_models.py
@@ -18,9 +18,10 @@ from idaes.core import FlowsheetBlock
 from idaes.generic_models.properties.core.generic.generic_property import GenericParameterBlock
 from idaes.core.util.scaling import calculate_scaling_factors
 from watertap.property_models import seawater_prop_pack
-from watertap.flowsheets.full_treatment_train.model_components import seawater_salt_prop_pack, seawater_ion_prop_pack
-from watertap.flowsheets.full_treatment_train.model_components.eNRTL import entrl_config_FpcTP
-from watertap.flowsheets.full_treatment_train.util import solve_with_user_scaling
+from watertap.examples.flowsheets.full_treatment_train.model_components import seawater_ion_prop_pack, \
+    seawater_salt_prop_pack
+from watertap.examples.flowsheets.full_treatment_train.model_components.eNRTL import entrl_config_FpcTP
+from watertap.examples.flowsheets.full_treatment_train.util import solve_with_user_scaling
 
 
 def build_prop(m, base='TDS'):
