+++ watertap/examples/flowsheets/full_treatment_train/model_components/tests/test_prop_pack.py
@@ -13,14 +13,14 @@
 
 import pytest
 from watertap.property_models.tests.property_test_harness import \
-    (PropertyTestHarness, PropertyRegressionTest)
+    (PropertyTestHarness)
 from pyomo.environ import ConcreteModel, value
 from idaes.core import FlowsheetBlock
 import idaes.core.util.scaling as iscale
 from pyomo.util.check_units import assert_units_consistent
-from watertap.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
-import watertap.flowsheets.full_treatment_train.model_components.seawater_ion_prop_pack as property_seawater_ions
-import watertap.flowsheets.full_treatment_train.model_components.seawater_salt_prop_pack as property_seawater_salts
+from watertap.examples.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
+import watertap.examples.flowsheets.full_treatment_train.model_components.seawater_ion_prop_pack as property_seawater_ions
+import watertap.examples.flowsheets.full_treatment_train.model_components.seawater_salt_prop_pack as property_seawater_salts
 
 # -----------------------------------------------------------------------------
 @pytest.mark.component
