+++ watertap/examples/flowsheets/full_treatment_train/model_components/tests/test_property_models.py
@@ -12,7 +12,7 @@
 ###############################################################################
 import pytest
 from pyomo.environ import value
-from watertap.flowsheets.full_treatment_train.model_components import property_models
+from watertap.examples.flowsheets.full_treatment_train.model_components import property_models
 
 
 @pytest.mark.component
