+++ watertap/examples/flowsheets/full_treatment_train/model_components/unit_0DRO.py
@@ -11,22 +11,22 @@
 #
 ###############################################################################
 
-"""1D reverse osmosis examples"""
+"""0D reverse osmosis examples"""
 
 from pyomo.environ import ConcreteModel
 from idaes.core import FlowsheetBlock
 from idaes.core.util.scaling import calculate_scaling_factors
-from watertap.flowsheets.full_treatment_train.model_components import property_models
-from watertap.unit_models.reverse_osmosis_1D import (ReverseOsmosis1D,
+from watertap.examples.flowsheets.full_treatment_train.model_components import property_models
+from watertap.unit_models.reverse_osmosis_0D import (ReverseOsmosis0D,
                                                        ConcentrationPolarizationType,
                                                        MassTransferCoefficient,
                                                        PressureChangeType)
-from watertap.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
+from watertap.examples.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
 
 
 def build_RO(m, base='TDS', level='simple', name_str='RO'):
     """
-    Builds a 1DRO model at a specified level (simple or detailed).
+    Builds a 0DRO model at a specified level (simple or detailed).
     Requires prop_TDS property package.
     """
     if base not in ['TDS']:
@@ -35,31 +35,37 @@ def build_RO(m, base='TDS', level='simple', name_str='RO'):
     prop = property_models.get_prop(m, base=base)
 
     if level == 'simple':
-        raise ValueError('Unexpected RO level {level} for build_RO'
-                         ''.format(level=level))
-
-    elif level == 'detailed':
         # build unit
-        setattr(m.fs, name_str, ReverseOsmosis1D(default={
+        setattr(m.fs, name_str, ReverseOsmosis0D(default={
             "property_package": prop,
-            "has_pressure_change": True,
-            "pressure_change_type": PressureChangeType.calculated,
-            "mass_transfer_coefficient": MassTransferCoefficient.calculated,
-            "concentration_polarization_type": ConcentrationPolarizationType.calculated,
-            "transformation_scheme": "BACKWARD",
-            "transformation_method": "dae.finite_difference",
-            "finite_elements": 10,
-        }))
+            "mass_transfer_coefficient": MassTransferCoefficient.none,
+            "concentration_polarization_type": ConcentrationPolarizationType.none}))
         blk = getattr(m.fs, name_str)
 
         # specify unit
         blk.area.fix(50)
         blk.A_comp.fix(4.2e-12)
         blk.B_comp.fix(3.5e-8)
-        blk.mixed_permeate[0].pressure.fix(101325)
+        blk.permeate.pressure[0].fix(101325)
+
+    elif level == 'detailed':
+        # build unit
+        setattr(m.fs, name_str, ReverseOsmosis0D(default={
+            "property_package": prop,
+            "has_pressure_change": True,
+            "pressure_change_type": PressureChangeType.calculated,
+            "mass_transfer_coefficient": MassTransferCoefficient.calculated,
+            "concentration_polarization_type": ConcentrationPolarizationType.calculated}))
+        blk = getattr(m.fs, name_str)
+
+        # specify unit
+        blk.area.fix(50)
+        blk.A_comp.fix(4.2e-12)
+        blk.B_comp.fix(3.5e-8)
+        blk.permeate.pressure[0].fix(101325)
         blk.channel_height.fix(1e-3)
         blk.spacer_porosity.fix(0.97)
-        blk.N_Re[0, 0].fix(500)
+        blk.N_Re_io[0, 'in'].fix(500)
 
     else:
         raise ValueError('Unexpected argument {level} for level in build_RO'
@@ -74,8 +80,8 @@ def solve_RO(base='TDS', level='simple'):
     build_RO(m, base=base, level=level)
 
     # specify feed
-    property_models.specify_feed(m.fs.RO.feed_side.properties[0, 0], base='TDS')
-    m.fs.RO.feed_side.properties[0, 0].pressure.fix(50e5)
+    property_models.specify_feed(m.fs.RO.feed_side.properties_in[0], base='TDS')
+    m.fs.RO.feed_side.properties_in[0].pressure.fix(50e5)
 
     # scaling
     calculate_scaling_factors(m)
@@ -83,7 +89,6 @@ def solve_RO(base='TDS', level='simple'):
     # initialize
     m.fs.RO.initialize(optarg={'nlp_scaling_method': 'user-scaling'})
 
-    m.fs.RO.display()
     check_dof(m)
     solve_with_user_scaling(m)
 
@@ -93,5 +98,5 @@ def solve_RO(base='TDS', level='simple'):
 
 
 if __name__ == "__main__":
-    # solve_RO(base='TDS', level='simple')
+    solve_RO(base='TDS', level='simple')
     solve_RO(base='TDS', level='detailed')
