+++ watertap/flowsheets/full_treatment_train/flowsheet_components/chemistry/SepRO_plus_Chlorination.py
@@ -47,7 +47,7 @@
     ---------- NOTE: This is only an example ---------
 """
 
-from watertap.flowsheets.full_treatment_train.model_components import unit_separator, property_models
+from watertap.examples.flowsheets.full_treatment_train.model_components import unit_separator, property_models
 from watertap.flowsheets.full_treatment_train.flowsheet_components.chemistry.PostTreatment_SimpleNaOCl_Chlorination import (
     build_simple_naocl_chlorination_unit,
     initialize_chlorination_example,
@@ -65,13 +65,13 @@ from idaes.core import FlowsheetBlock
 from idaes.generic_models.unit_models.translator import Translator
 from pyomo.network import Arc
 
-from watertap.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
+from watertap.examples.flowsheets.full_treatment_train.util import solve_with_user_scaling, check_dof
 
-from watertap.flowsheets.full_treatment_train.electrolyte_scaling_utils import (
+from watertap.examples.flowsheets.full_treatment_train.electrolyte_scaling_utils import (
     approximate_chemical_state_args,
     calculate_chemical_scaling_factors)
 
-from watertap.flowsheets.full_treatment_train.chemical_flowsheet_util import seq_decomp_initializer
+from watertap.examples.flowsheets.full_treatment_train.chemical_flowsheet_util import seq_decomp_initializer
 
 from idaes.core.util import scaling as iscale
