+++ watertap/flowsheets/full_treatment_train/flowsheet_components/costing.py
@@ -12,22 +12,11 @@
 ###############################################################################
 
 from pyomo.environ import (
-    Block, ConcreteModel, Constraint, Expression, Var, Param, value, TransformationFactory, units as pyunits)
+    Block, Constraint, Expression, Var, Param, value)
 import watertap.flowsheets.full_treatment_train.flowsheet_components.financials as financials
-from watertap.flowsheets.full_treatment_train.flowsheet_components import feed_block
-from watertap.flowsheets.full_treatment_train.model_components import unit_separator, unit_0DRO, unit_1DRO, property_models
 
-from watertap.flowsheets.full_treatment_train.flowsheet_components.desalination import (build_desalination,
-                                                                                          solve_desalination,
-                                                                                          scale_desalination,
-                                                                                          initialize_desalination,
-                                                                                          display_desalination)
 # from watertap.flowsheets.full_treatment_train.analysis.flowsheet_softening_two_stage import build, scale, initialize
-from watertap.flowsheets.full_treatment_train.util import (solve_with_user_scaling,
-                                                             check_dof)
 import idaes.core.util.scaling as iscale
-from idaes.core import FlowsheetBlock
-
 
 cost_capacity_flag = False
