+++ watertap/flowsheets/full_treatment_train/flowsheet_components/feed_block.py
@@ -13,10 +13,8 @@
 
 """Feed blocks for supported property packages"""
 
-from pyomo.environ import Constraint
 from idaes.generic_models.unit_models import Feed
-from watertap.flowsheets.full_treatment_train.model_components import property_models
-from idaes.core.util.scaling import calculate_scaling_factors
+from watertap.examples.flowsheets.full_treatment_train.model_components import property_models
 
 
 def build_feed(m, base='TDS'):
