+++ watertap/flowsheets/full_treatment_train/flowsheet_components/tests/test_pretreatment_softening.py
@@ -14,8 +14,7 @@ import pytest
 from pyomo.environ import ConcreteModel, value, TransformationFactory
 from idaes.core import FlowsheetBlock
 from watertap.flowsheets.full_treatment_train.flowsheet_components import pretreatment_softening
-from watertap.flowsheets.full_treatment_train.model_components import property_models
-from watertap.flowsheets.full_treatment_train.util import check_build, check_dof, check_scaling
+from watertap.examples.flowsheets.full_treatment_train.util import check_dof
 
 
 @pytest.mark.component
